Used IntelliJ's "Code Inspection" to look for and automatically fix issues.  It reported many potential problems and suggestions, some of which were obvious improvements.  Fixed those (see below).  There should be no change to Sphinx4's behavior expect perhaps improved speed performance.
Generify more code

e.g.
        for (Iterator i = map.keySet().iterator(); i.hasNext();) {
            WordSequence ws = (WordSequence)i.next();           
            |||
            VVV           
        for (Iterator<WordSequence> i = map.keySet().iterator(); i.hasNext();) {
            WordSequence ws = i.next();

Remove unnecessary boxing and unboxing

e.g.

	samples.add(new Short((short) val)); --> samples.add((short) val);



Remove unnecessary casts        
        
        double rootMeanSquare = Math.sqrt
            ((double)sumOfSquares/samples.length);                   
            |||
            VVV                                  
        double rootMeanSquare = Math.sqrt
            (sumOfSquares /samples.length);
            

Remove unnecessary semicolon

	e.g.
	
		foo();; --> foo();

        
Added missing CloneNotSupportedException to clone() methods

e.g.
	public Object clone() {
            |||
            VVV                        
	public Object clone()throws CloneNotSupportedException {
	

Removed unnecessary bit operations

e.g. 

	int high = (int) ((id >> 32) & 0xffffffff); --> int high = (int) (id >> 32);
	
	
Using System.arraycopy() to improve performance

e.g.
	
        for (int i = 0; i < outSampleIndex; i++) {
            outSamples[i] = temp[i];
        }                  
            |||
            VVV           
        System.arraycopy(temp, 0, outSamples, 0, outSampleIndex);
        
Replacing String.append() with StringBuffer.append() to improve performance

e.g. 
        
        sb.append("s: " + startTime + " e: " + endTime + "[");                   
            |||
            VVV                        
        sb.append("s: ").append(startTime).append(" e: ").append(endTime).append("[");

git-svn-id: svn+ssh://svn.code.sf.net/p/cmusphinx/code/trunk/sphinx4@6329 94700074-3cef-4d97-a70e-9c8c206c02f5

+++ src/sphinx4/edu/cmu/sphinx/frontend/endpoint/SpeechClassifier.java
@@ -189,7 +189,7 @@ public class SpeechClassifier extends BaseDataProcessor {
             sumOfSquares += sample * sample;
         }
         double rootMeanSquare = Math.sqrt
-            ((double)sumOfSquares/samples.length);
+            (sumOfSquares /samples.length);
         rootMeanSquare = Math.max(rootMeanSquare, 1);
         return (LogMath.log10((float)rootMeanSquare) * 20);
     }
